backend/pyproject.toml
[project]
name = "upguardian-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.121.1",
    "uvicorn>=0.38.0",
    "PyJWT[crypto]>=2.8.0",
    "requests>=2.30.0",
    "python-dotenv>=1.2.1",
    "cryptography>=46.0.3",
    "dotenv>=0.9.9",
    "openai>=2.7.1",
]

[project.scripts]
upguardian-backend = "upguardian_backend:main"

[build-system]
requires = ["uv_build>=0.9.8,<0.10.0"]
build-backend = "uv_build"
